requirements.txt
asgiref==3.11.0
boto3==1.42.1
botocore==1.42.1
crispy-bootstrap4==2025.6
dj-database-url==3.0.1
Django==5.2.9
django-allauth==65.13.1
django-countries==8.2.0
django-crispy-forms==2.5
django-environ==0.12.0
django-storages==1.14.6
jmespath==1.0.1
pillow==12.0.0
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
